+++ docs/CHANGE_LOG.md
@@ -9,8 +9,11 @@
 - Added a new binding option called "tooltipDelay", which states how long the tooltip should wait (in milliseconds) until it's shown (defaults to 1000).
 
 #### **Public Functions:**
-- Added new public function "addDates()", which will add an array of dates (saves all addDate() each time.)
-- Added new public function "removeDates()", which will remove an array of dates (saves all removeDate() each time.)
+- Added new public function "addDates()", which will add an array of dates (saves all addDate() each time).
+- Added new public function "removeDates()", which will remove an array of dates (saves all removeDate() each time).
+
+#### **CSS:**
+- Added an extra bit of spacing before the first bar in the chart view.
 
 <br>